ultimos cambios con grafico

+++ TP1/TP1_Sim_GrupoE/Clases/Frecuencia.cs
@@ -11,7 +11,7 @@ namespace TP1_Sim_GrupoE.Clases
         public int indice { get; set; }
         public Double LimInferior { get; set; }
         public Double LimSuperior { get; set; }
-        public int FrecObservada { get; set; }
+        public Double FrecObservada { get; set; }
         public Double FrecEsperado { get; set; }
         public Double Chi { get; set; }
         public Double SumChi { get; set; }
